requirements.txt
astropy==7.0.0
astropy_healpix==1.0.3
gwpy==3.0.11
h5py==3.12.1
healpy==1.18.1
hpmoc==1.0.0
ligo_gracedb==2.13.0
ligo_segments==1.4.0
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
pydantic==2.11.6
python_ligo_lw==1.8.3
scipy==1.15.3
tqdm==4.66.5
